requirements/local.txt
asgiref==3.3.4
Django==3.2.3
django-cors-headers==4.0.0
djangorestframework==3.12.4
greenlet==2.0.2
hdbcli==2.16.26
msgpack==1.0.5
mysqlclient==2.0.3
PyMySQL==1.0.2
pynvim==0.4.3
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
Unipath==1.1
